Update eval to avoid some problems
Previously, this code `eval`ed the contents of `config.ru` in a single
line between `{ ... }` and without specifying the optional arguments to
`Kernel.eval`.

I noticed three problems that this caused:

1. if the final line in `config.ru` was a comment, it effectively
   commented out the closing `}`, turning the `eval` into a syntax
   error
2. It was not possible to use `require_relative` in `config.ru`
3. Error messages caused by exceptions in `config.ru` did not present
   a correct stack trace

These problems are addressed in this commit by:

1. Changing the body of the eval to call `Rack::Builder.new` with a
   multiline block.
2. Passing the optional arguments to eval with the original filename
   of the `config.ru` and a correct line number.

This commit also passes `TOPLEVEL_BINDING` as the optional second
parameter. I didn't experience any specific problems caused by leaving
this out, but `eval` without an explicit binding would have caused the
local variables `config_ru` and others to be in scope of the user's
`config.ru` which seems undesirable and pointless.